Factor out the parser state from the Compiler
Inorder to share the parser state between multiple
instances of the compiler we needed the parser state
to outlive the liftime of the top-level
Compiler object. And also becasue we don't like global
variables. The only way a Compiler is constructed
is by giving it an instance of the parser-state
by reference. In this case the Virtual Machine
is the owner of the parser state.1