Emit OP_CLOSURE operands.
With this commit we now emit the operands required
by OP_CLOSURE.

These include the index into the function object in the constant pool.
We also emit all the upvalue metadata. Specifically the information
whether the upvalue refers to a local of an enclosing scope or
not. We also append the upvalue index.